requirements.txt
aiohttp
beautifulsoup4
python-dotenv
beautifulsoup4
aiogram